dom/power/test/test_power_set_key_light_enabled.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test Enabling/Disabling Screen with Power Management API</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
  <script type="application/javascript">

  "use strict";

  function testEnableKeyLight() {
    try {
      navigator.mozPower.keyLightEnabled = true;
      ok(navigator.mozPower.keyLightEnabled === true, "Enabled key backlight.");
    } catch (e) {
      ok(false, "Unexpected exception trying to enable key backlight.");
    }
  }

  function testDisableKeyLight() {
    try {
      navigator.mozPower.keyLightEnabled = false;
      ok(navigator.mozPower.keyLightEnabled === false, "Disabled key backlight.");
    } catch (e) {
      ok(false, "Unexpected exception trying to disable key backlight.");
    }
  }

  function startTests() {

    // Make sure we don't suspend
    navigator.mozPower.cpuSleepAllowed = false;

    testDisableKeyLight();
    testEnableKeyLight();

    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  if (SpecialPowers.hasPermission("power", document)) {
    // Currently only applicable on FxOS
    if (navigator.userAgent.indexOf("Mobile") != -1 &&
        navigator.appVersion.indexOf("Android") == -1) {
      startTests();
    } else {
      ok(true, "mozPower on Firefox OS only.");
      SimpleTest.finish();
    }
  } else {
    // Add the permission and reload so it's propogated
    SpecialPowers.addPermission("power", true, document);
    window.location.reload();
  }
  </script>
</pre>
</body>
</html>
